Separate snapshot construction from resolution coordination
Extract source and library snapshot construction while keeping pending work, clocks, and publication in the coordinator. Inject request policy and use production scheduling in tests.

Keep temporary project unavailability retryable through PSI classification and fallback. The lifecycle regression fails under the previous terminal behavior. Verified all 255 focused snapshot, service, module, highlighting, and navigation tests.
